refactor: feedback api 슬라이스 분리 (#67)
