+++ src/entities/feedback/index.ts
@@ -1,0 +1,2 @@
+export { feedbackPayloadSchema } from './model/feedback-schema'
+export { FeedbackApi } from './api/feedback-service'
