+++ src/entities/feedback/model/feedback-interface.ts
@@ -1,0 +1,5 @@
+import { FeedbackPayload } from './feedback-schema'
+
+export interface FeedbackService {
+  sendReportMail: (payload: FeedbackPayload) => void
+}
