+++ src/entities/speller/index.ts
@@ -6,12 +6,10 @@ export {
   CorrectMethodEnum,
   checkPayloadSchema,
   checkResponseSchema,
-  feedbackPayloadSchema,
   type ErrorInfo,
   type CorrectInfo,
   type CheckPayload,
   type CheckResponse,
-  type FeedbackPayload,
 } from './model/speller-schema'
 export { SpellerApi } from './api/speller-service'
 export { useSpeller } from './model/use-speller'
